chore: extract pkg contacts
