+++ pkg/contacts/contact.go
@@ -1,4 +1,4 @@
-package protocol
+package contacts
 
 import (
 	"crypto/ecdsa"
