+++ pkg/contacts/contact_test.go
@@ -1,4 +1,4 @@
-package protocol
+package contacts
 
 import (
 	"encoding/json"
